refactor(server): consolidate server files under server/ and remove language logic duplication
Move server.mjs → server/index.mjs so all server-side code lives in one
place alongside public-language-routing.mjs and its tests.

server/index.mjs now imports resolvePublicLanguageRedirect and
getSupportedPublicLanguagesFromEnv from the sibling routing module
instead of inlining the same logic verbatim. The routing module was
already extracted and tested; server.mjs just never imported from it.

Also introduce a local `root` constant (one level up from __dirname)
so dist/ and Vike dev-middleware paths remain correct after the move.

Update serve:ssr script to point at server/index.mjs.
