+++ Dockerfile
@@ -18,7 +18,7 @@ ENV PORT=3000
 COPY --from=builder /app/package.json ./package.json
 COPY --from=builder /app/node_modules ./node_modules
 COPY --from=builder /app/dist ./dist
-COPY --from=builder /app/server.mjs ./server.mjs
+COPY --from=builder /app/server ./server
 
 EXPOSE 3000
-CMD ["node", "server.mjs"]
+CMD ["node", "server/index.mjs"]
